### 2023-07-23 #5634 @arfon
:wave: @idoby @jkbren@ nidhogwould any of you be willing to review this submission for JOSS? The submission under consideration is _PyDGN: a Python Library for Flexible and Reproducible Research on Deep Learning for Graphs_

Many thanks
Arfon
