### 2023-07-27 f29600b
removed files to make PyDGN work with Python 3.10
