### 2023-07-27 98dfd63
removed files to make PyDGN work with Python 3.10
