+++ .github/requirements.txt
@@ -1,4 +1,4 @@
-PyYAML==5.4
+PyYAML==6.0
 requests==2.22.0
 tensorboard==2.11.0
 tqdm==4.47.0
