### 2023-07-27 a973dc1
using Ray 2.6.0 to support Python 3.10
