+++ .github/requirements.txt
@@ -5,7 +5,7 @@ tqdm==4.47.0
 ogb==1.3.5
 protobuf==3.20.3
 click<=8.0.4
-ray==2.1.0
+ray==2.6.0
 gpustat==1.0.0
 wandb==0.12.15
 torch==1.13.0
