+++ setup.py
@@ -37,8 +37,8 @@ setuptools.setup(
         "ray == 2.6.0",
         "gpustat >= 1.0.0",
         "torch <= 1.13.0",
-        "torch-geometric <= 2.1.0.post1",
-        "torch-geometric-temporal <= 0.54.0",
+        "torch-geometric == 2.3.0",
+        "torch-geometric-temporal == 0.54.0",
         "wandb >= 0.12.15",
     ],
     packages=setuptools.find_packages(),
