### 2023-07-27 6ad58f2
Update python-test-and-coverage.yml
Reverting erroneous change that should have been done in a separate branch.
