+++ setup.py
@@ -37,7 +37,7 @@ setuptools.setup(
         "ray == 2.6.0",
         "gpustat >= 1.0.0",
         "torch <= 1.13.0",
-        "torch-geometric == 2.3.0",
+        "torch-geometric <= 2.3.0",
         "torch-geometric-temporal == 0.54.0",
         "wandb >= 0.12.15",
     ],
