+++ .github/requirements.txt
@@ -1,5 +1,5 @@
 PyYAML==6.0
-requests==2.22.0
+requests==2.31.0
 tensorboard==2.11.0
 tqdm==4.47.0
 ogb==1.3.5
