+++ setup.py
@@ -1,6 +1,6 @@
 import setuptools
 
-__version__ = "1.3.1"
+__version__ = "1.4.0"
 
 URL = "https://pydgn.readthedocs.io/en/latest/"
 
