+++ docs/conf.py
@@ -24,7 +24,7 @@ copyright = "2022, Federico Errica"
 author = "Federico Errica"
 
 # The full version, including alpha/beta/rc tags
-release = "1.3.1"
+release = "1.4.0"
 
 
 # -- General configuration ---------------------------------------------------
