+++ docs/conf.py
@@ -24,7 +24,7 @@ copyright = "2022, Federico Errica"
 author = "Federico Errica"
 
 # The full version, including alpha/beta/rc tags
-release = "1.4.0"
+release = "1.4.1"
 
 
 # -- General configuration ---------------------------------------------------
