+++ .github/workflows/interrogate-docstring.yaml
@@ -14,10 +14,10 @@ jobs:
     steps:
     - uses: actions/checkout@v2
 
-    - name: Set up Python 3.10
-      uses: actions/setup-python@v3
+    - name: Set up Python 3.8
+      uses: actions/setup-python@v2
       with:
-        python-version: 3.10
+        python-version: 3.8
 
     - name: Python Interrogate Check
       uses: diningphil/python-interrogate-check@main
