### 2023-07-30 2900ae0
added new utility for retrieving and filtering configurations
