+++ setup/create_environment.sh
@@ -1,7 +1,7 @@
 # default pytorch version
-PYTHON=python3.9
+PYTHON=python3.8
 PYTORCH_VERSION=1.13.0
-PYTORCH_GEOMETRIC_VERSION=2.3.0
+PYTORCH_GEOMETRIC_VERSION=2.3.0  #  since 2.3.0, all other packages are not anymore necessary
 
 # set CUDA variable (defaults to cpu if no argument is provided to the script)
 CUDA_VERSION=${1:-cpu}
@@ -15,18 +15,19 @@ pip install build wheel pytest black
 
 # install pytorch
 if [[ "$CUDA_VERSION" == "cpu" ]]; then
-  pip3 install torch==${PYTORCH_VERSION} torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cpu
+  pip install torch==${PYTORCH_VERSION} torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cpu
 elif [[ "$CUDA_VERSION" == 'cu112' ]]; then
-  pip3 install torch==${PYTORCH_VERSION} torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu112
+  pip install torch==${PYTORCH_VERSION} torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu112
 elif [[ "$CUDA_VERSION" == 'cu113' ]]; then
-  pip3 install torch==${PYTORCH_VERSION} torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu113
+  pip install torch==${PYTORCH_VERSION} torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu113
 elif [[ "$CUDA_VERSION" == 'cu116' ]]; then
-  pip3 install torch==${PYTORCH_VERSION} torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu116
+  pip install torch==${PYTORCH_VERSION} torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu116
 elif [[ "$CUDA_VERSION" == 'cu117' ]]; then
-  pip3 install torch==${PYTORCH_VERSION} torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu117
+  pip install torch==${PYTORCH_VERSION} torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu117
 fi
 
-pip install pyg-lib torch-scatter torch-sparse torch-cluster torch-spline-conv torch-geometric==${PYTORCH_GEOMETRIC_VERSION} -f https://data.pyg.org/whl/torch-${PYTORCH_VERSION}+${CUDA_VERSION}.html
+# pip install pyg-lib torch-scatter torch-sparse torch-cluster torch-spline-conv torch-geometric==${PYTORCH_GEOMETRIC_VERSION} -f https://data.pyg.org/whl/torch-${PYTORCH_VERSION}+${CUDA_VERSION}.html
+pip install torch-geometric==${PYTORCH_GEOMETRIC_VERSION}
 
 pip install jupyter
 pip install pydgn
