+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.4.2] 
+
+### Added
+
+- Utility functions in `pydgn.evaluation.util` to retrieve configuration files in a model selection folder and filter them for post-hoc analyses. 
+
 ## [1.4.1] Support for Python 3.8-3.11
 
 ### Changed
