+++ setup.py
@@ -1,6 +1,6 @@
 import setuptools
 
-__version__ = "1.4.1"
+__version__ = "1.4.2"
 
 URL = "https://pydgn.readthedocs.io/en/latest/"
 
