+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Utility functions in `pydgn.evaluation.util` to retrieve configuration files in a model selection folder and filter them for post-hoc analyses. 
+- Updated tutorial in documentation accordingly
 
 ## [1.4.1] Support for Python 3.8-3.11
 
