### 2023-07-31 #5634 @arfon
:wave: @sepandhaghighi @matteocao – would either of you be willing to review this submission for JOSS? The submission under consideration is PyDGN: a Python Library for Flexible and Reproducible Research on Deep Learning for Graphs.

Many thanks
Arfon
