### 2023-08-01 #5713 @editorialbot
[REVIEW]: PyDGN: a Python Library for Flexible and Reproducible Research on Deep Learning for Graphs
**Submitting author:** @diningphil (<a href="http://orcid.org/0000-0001-5181-2904">Federico Errica</a>)
**Repository:** https://github.com/diningphil/PyDGN/
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.5.0
