### 2023-08-01 #5713 @arfon
@sepandhaghighi – looks like there's a whitespace character at the start of that command to @editorialbot which is stopping it from responding (a bug we should fix).

Could you retry in a new comment, ensuring the sentence starts with `@editorialbot`.
