### 2023-08-21 #5713 @idoby
> 👋 @idoby & @sepandhaghighi – just checking in to see how you're getting on with your reviews? It looks like you've both made a start here, do you think you might be able to wrap up your initial reviews in the next week or so so that @diningphil can start responding?

Thanks for the reminder, had forgotten about this. Will comment soon
