### 2023-09-06 27e33a8
improvements after JOSS review and minor stuff
