+++ .github/workflows/python-test-and-coverage.yml
@@ -31,8 +31,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install flake8 pytest coverage coverage-badge
-        pip install -r .github/requirements.txt
-        pip install -r .github/requirements_torch_geometric.txt
+        python -m pip install .
 
     - name: Lint with flake8
       run: |
@@ -43,15 +42,15 @@ jobs:
       run: |
         coverage run -m pytest tests/
         coverage report -m
-        rm badges/coverage_badge.svg
-        coverage-badge -o badges/coverage_badge.svg
+        rm .badges/coverage_badge.svg
+        coverage-badge -o .badges/coverage_badge.svg
 
     - name: Commit files
       if: github.event_name == 'push'
       run: |
         git config --local user.email "action@github.com"
         git config --local user.name "GitHub Action"
-        git add badges/coverage_badge.svg
+        git add .badges/coverage_badge.svg
         git diff --exit-code || git commit -m "Update coverage_badge.svg" -a
 
     - name: Push changes
