+++ .github/workflows/interrogate-docstring.yaml
@@ -12,18 +12,18 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v3
 
-    - name: Set up Python 3.8
-      uses: actions/setup-python@v2
+    - name: Set up Python 3.11
+      uses: actions/setup-python@v3
       with:
-        python-version: 3.8
+        python-version: 3.11
 
     - name: Python Interrogate Check
       uses: diningphil/python-interrogate-check@main
       with:
         path: 'pydgn'
-        badge-location: 'badges/interrogate_badge.svg'
+        badge-location: '.badges/interrogate_badge.svg'
 
 
     - name: Commit files
