+++ paper.md
@@ -32,13 +32,13 @@ bibliography: paper.bib
 
 # Summary
 
-The disregard for standardized evaluation procedures is a longstanding issue in the Machine Learning (ML) community that severely hampers real progress. This is especially true for fast-growing research areas, where a substantial amount of literature relentlessly appears every day. In the graph machine learning field, these issues have already been brought to light and partially addressed, but a general-purpose library for rigorous evaluations and reproducible experiments is lacking in the graph machine learning landscape. We therefore introduce a new Python library, called `PyDGN`, to provide users with a system that lets them focus on models' development while ensuring empirical rigor and reproducibility of their results.
+The use of standardized evaluation procedures is a key component in the Machine Learning (ML) field to determine whether new approaches grant real advantages over others. This is especially true for fast-growing research areas, where a substantial amount of literature relentlessly appears every day. In the graph machine learning field, some evaluation issues have already been brought to light and partially addressed, but a general-purpose library for rigorous evaluations and reproducible experiments is lacking in the graph machine learning landscape. We therefore introduce a new Python library, called `PyDGN`, to provide users with a system that lets them focus on models' development while ensuring empirical rigor and reproducibility of their results.
 
 # Statement of need
 
-To date, the graph ML community [@1997sperduti_supervised_1997; @2009scarselli_graph_2009; @2009micheli_neural_2009; @2017bronstein_geometric_2017; @2017hamilton_representation_2017; @2020wu_comprehensive_2020] has already developed benchmarking software to re-evaluate existing models on a fixed set of datasets [@2018shchur_pitfalls_2018; @2020errica_fair_2020; @2020hu_open_2020; @2021liu_dig_2021]. In addition, existing libraries such as Pytorch Geometric (PyG) [@fey_fast_2019], Deep Graph Library (DGL) [@2019wang_dgl_2019], and Spektral [@2021grattarola_graph_2021] provide the building blocks of Deep Graph Networks (DGNs) [@2020bacciu_gentle_2020], also known as message-passing architectures [@2017gilmer_neural_2017], effectively acting as the backbone of most graph ML software packages.
+To date, the graph ML community [@1997sperduti_supervised_1997; @2009scarselli_graph_2009; @2009micheli_neural_2009; @2017bronstein_geometric_2017; @2017hamilton_representation_2017; @2020wu_comprehensive_2020] has already developed benchmarking software to re-evaluate existing models on a fixed set of datasets [@2018shchur_pitfalls_2018; @2020errica_fair_2020; @2020hu_open_2020; @2021liu_dig_2021]. In addition, existing libraries such as PyTorch Geometric (PyG) [@2019fey_fast_2019], Deep Graph Library (DGL) [@2019wang_dgl_2019], and Spektral [@2021grattarola_graph_2021] provide the building blocks of Deep Graph Networks (DGNs) [@2020bacciu_gentle_2020], also known as message-passing architectures [@2017gilmer_neural_2017], effectively acting as the backbone of most graph ML software packages. Other libraries, such as GraphGym [@2020_you_design_2020], allow the user to explore the design of existing DGNs by running hyper-parameter tuning experiments in parallel, but the customization is mostly limited to the models and does not allow, for instance, a modular extension of data splitting techniques, evaluation strategies, or experiments with a custom logic. This limits the ability of a researcher to use these libraries to carry out new and possibly original research.  
 
[…]
