+++ paper.bib
@@ -7,6 +7,7 @@
 	year = {1997},
 	publisher={IEEE},
 	pages = {714--735}
+    doi={10.1109/72.572108}
 }
 
 @article{2009scarselli_graph_2009,
@@ -17,7 +18,8 @@
 	author = {Scarselli, Franco and Gori, Marco and Tsoi, Ah Chung and Hagenbuchner, Markus and Monfardini, Gabriele},
 	year = {2009},
 	publisher={IEEE},
-	pages = {61--80}
+	pages = {61--80},
+	doi={10.1109/TNN.2008.2005605}
 }
 
 @article{2009micheli_neural_2009,
@@ -28,7 +30,9 @@
 	author = {Micheli, Alessio},
 	year = {2009},
 	publisher={IEEE},
-	pages = {498--511}
+	pages = {498--511},
+    doi={10.1109/TNN.2008.2010350}
+
 }
 
 
@@ -39,7 +43,8 @@
 	journal = {IEEE Signal Processing Magazine},
 	author = {Bronstein, Michael M. and Bruna, Joan and LeCun, Yann and Szlam, Arthur and Vandergheynst, Pierre},
 	year = {2017},
-	pages = {25. 18--42}
+	pages = {25. 18--42},
+    doi={10.1109/MSP.2017.2693418}
 }
 
 @inproceedings{2017gilmer_neural_2017,
@@ -89,6 +94,13 @@
   year={2019}
 }
 
+@inproceedings{2020_you_design_2020,
+  title={Design space for graph neural networks},
+  author={You, Jiaxuan and Ying, Zhitao and Leskovec, Jure},
+  booktitle={Proceedings of the 34th Conference on Neural Information Processing Systems (NeurIPS)},
+  year={2020}
+}
+
 @article{2019wang_dgl_2019,
     title={Deep Graph Library: A Graph-Centric, Highly-Performant Package for Graph Neural Networks},
     author={Minjie Wang and Da Zheng and Zihao Ye and Quan Gan and Mufei Li and Xiang Song and Jinjing Zhou and Chao Ma and Lingfan Yu and Yu Gai and Tianjun Xiao and Tong He and George Karypis and Jinyang Li and Zheng Zhang},
@@ -104,7 +116,8 @@
 	author = {Zhang, Si and Tong, Hanghang and Xu, Jiejun and Maciejewski, Ross},
 	year = {2019},
 	publisher={Springer},
-	pages = {11}
+	pages = {11},
+	doi={10.1186/s40649-019-0069-y}
 }
 
 
@@ -116,7 +129,8 @@
   month={9},
   pages={203--221},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.neunet.2020.06.006}
 }
 
 @inproceedings{2020errica_fair_2020,
@@ -139,7 +153,8 @@
   author={Wu, Zonghan and Pan, Shirui and Chen, Fengwen and Long, Guodong and Zhang, Chengqi and Philip, S Yu},
   journal={IEEE Transactions on Neural Networks and Learning Systems},
   year={2020},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TNNLS.2020.2978386}
 }
 
 @article{2020zhang_deep_2020,
@@ -147,7 +162,8 @@
   author={Zhang, Ziwei and Cui, Peng and Zhu, Wenwu},
   journal={IEEE Transactions on Knowledge and Data Engineering},
   year={2020},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TKDE.2020.2981333}
 }
 
 @article{2021grattarola_graph_2021,
[…]
