+++ paper.bib
@@ -6,7 +6,7 @@
 	author = {Sperduti, Alessandro and Starita, Antonina},
 	year = {1997},
 	publisher={IEEE},
-	pages = {714--735}
+	pages = {714--735},
     doi={10.1109/72.572108}
 }
 
