+++ paper.md
@@ -38,7 +38,7 @@ The use of standardized evaluation procedures is a key component in the Machine 
 
 To date, the graph ML community [@1997sperduti_supervised_1997; @2009scarselli_graph_2009; @2009micheli_neural_2009; @2017bronstein_geometric_2017; @2017hamilton_representation_2017; @2020wu_comprehensive_2020] has already developed benchmarking software to re-evaluate existing models on a fixed set of datasets [@2018shchur_pitfalls_2018; @2020errica_fair_2020; @2020hu_open_2020; @2021liu_dig_2021]. In addition, existing libraries such as PyTorch Geometric (PyG) [@2019fey_fast_2019], Deep Graph Library (DGL) [@2019wang_dgl_2019], and Spektral [@2021grattarola_graph_2021] provide the building blocks of Deep Graph Networks (DGNs) [@2020bacciu_gentle_2020], also known as message-passing architectures [@2017gilmer_neural_2017], effectively acting as the backbone of most graph ML software packages. Other libraries, such as GraphGym [@2020_you_design_2020], allow the user to explore the design of existing DGNs by running hyper-parameter tuning experiments in parallel, but the customization is mostly limited to the models and does not allow, for instance, a modular extension of data splitting techniques, evaluation strategies, or experiments with a custom logic. This limits the ability of a researcher to use these libraries to carry out new and possibly original research.  
 
-In other words, the community lacks a software library that is specifically dedicated to ensuring reproducibility and replicability of experiments _without_ compromising the flexibility required by our everyday research. To fill this gap, we have developed `PyDGN`, a Python library for Deep Graph Networks research. `PyDGN` builds upon `PyTorch` [@2019paszke_PyTorch_2019] and `PyTorch Geometric` (PyG) [@2019fey_fast_2019] to handle graph-structured data and reuse efficient implementations of machine learning models. It exploits [Ray](https://www.ray.io/) to run experiments in parallel (also on clusters of machines) and it supports [GPU](https://developer.nvidia.com/cuda-toolkit) computation for faster executions. Our goal is to help practitioners and researchers to focus on the development of their models and to effortlessly evaluate them under fair, robust, and reproducible experimental conditions, thus mitigating empirical malpractices that often occur in the ML community [@2018lipton_troubling_2018; @2018shchur_pitfalls_2018; @2020errica_fair_2020]. `PyDGN` has already been used in a number of research projects that have been published at top-tier venues, as listed in the [official GitHub repository](https://github.com/diningphil/PyDGN).
+In other words, the community lacks a software library that is specifically dedicated to ensuring reproducibility and replicability of experiments _without_ compromising the flexibility required by our everyday research. To fill this gap, we have developed `PyDGN`, a Python library for Deep Graph Networks research. `PyDGN` builds upon `PyTorch` [@2019paszke_pytorch_2019] and `PyTorch Geometric` (PyG) [@2019fey_fast_2019] to handle graph-structured data and reuse efficient implementations of machine learning models. It exploits [Ray](https://www.ray.io/) to run experiments in parallel (also on clusters of machines) and it supports [GPU](https://developer.nvidia.com/cuda-toolkit) computation for faster executions. Our goal is to help practitioners and researchers to focus on the development of their models and to effortlessly evaluate them under fair, robust, and reproducible experimental conditions, thus mitigating empirical malpractices that often occur in the ML community [@2018lipton_troubling_2018; @2018shchur_pitfalls_2018; @2020errica_fair_2020]. `PyDGN` has already been used in a number of research projects that have been published at top-tier venues, as listed in the [official GitHub repository](https://github.com/diningphil/PyDGN).
 
[…]
