### 2023-09-08 #5713 @diningphil
Dear @idoby,

First of all, thank you for the constructive and positive feedback. We were able to simplify a lot the installation process thanks to your comments. Below you can find our response so that we can continue the discussion.

## Paper:

1. We added DOIs for all journal papers and the broken reference
2. You are right, we changed the sentence to make it more positive
3. We did our best to include snippets of code that give the reader an idea of what should be done to carry out the most important steps, namely the data splitting, which corresponds to the choice of an evaluation procedure such as 10-fold cross validation, the definition of a new model, and the definition of a new metric. Because all pieces of the puzzle come together in a way that is transparent to the user, we found it difficult to show other snippets, but we believe that this, combined to the [turorial](https://pydgn.readthedocs.io/en/latest/tutorial.html) in our documentation, might be enough to understand the overall mechanism. Any further suggestion is welcome. We also added a motivation when missing to the different paragraphs.
4. We added a new paragraph on key design considerations. We identified two of them, namely the use of configuration files to enable fast prototyping and the use of the publish-subscribe pattern to flexibly adapt components of the training loop without touching the main logic. Thank you very much for the suggestion!
5. We fixed this, thank you for noticing, altough  we cannot see any other inconsistency like that in the text.
6. We added the relevant citation and discussed how it differs from PyDGN. Please note that we already discussed to some extent how the other libraries mentioned are different from us: they are more focused on the definition of building blocks of graph machine learning models, whereas we try to automatize the experiments while keeping the framework flexible enough for everyday research. These libraries can be used in our framework, in particular Pytorch Geometric.

## Software (modified on main branch)

1. We completely agree and we simplified a lot the installation process, requiring only a `pip install pydgn` command (please refer to README.md in the **main** branch). We also specified the dependencies in the `toml` file and removed the legacy `setup` files as suggested. Only the strictly required dependencies are now listed. Thank you very much for the help, now the code looks much cleaner.
2. Handled in the previous point
3. Handled in the previous point

As a last note, please note that the example usage is shown in the readme file, and in the aforementioned tutorial the user can find an explanation of the configuration files and how to use them to setup a proper experiment. Examples of configuration files can be also found in the `examples` folder.

Thank you again for the help. We are happy to discuss any more suggestions, if needed, related to the new version of the code.
I'd also like to tag @arfon to show that the discussion is ongoing =).

Best regards, the authors!
