### 2023-09-09 #5713 @idoby
@diningphil I think the intro reads much better and is better motivated now, and the differences between the various existing packages and yours are clearer. The practical use and motivation for the features and design decisions reads well.

Note that something seems to have not rendered well in line 83 (seems like a header didn't render) and that PyDGN is styled inconsistently on line 95.

Regarding the installation procedure: PyDGN should install much more easily now. Please consider not forcing the user to install gpustat, since not all installs have CUDA. The same applies to the wandb dependency: please consider detecting wandb at runtime, not forcing your users to install wandb if they don't use it.

Besides that, I think we're good to go. Please consider updating the paper branch with the latest changes so that when the branch is archived upon acceptance, it would include the changes to the installation procedure and any other enhancement you made to the software.

Good job! 👍
