### 2023-09-09 #5713 @idoby
Hmm, this list doesn't seem to include all of the references in the paper...
