+++ .github/requirements.txt
@@ -1,11 +1,10 @@
-PyYAML==5.4
-requests==2.22.0
-tensorboard==2.11.0
-tqdm==4.47.0
-ogb==1.3.5
-protobuf==3.20.3
-click<=8.0.4
-ray==2.1.0
-gpustat==1.0.0
-wandb==0.12.15
-torch==1.13.0
+PyYAML
+requests
+tensorboard
+tqdm
+ogb
+ray
+gpustat
+wandb
+torch
+torch-geometric>=2.3.0
