+++ .github/requirements.txt
@@ -4,7 +4,5 @@ tensorboard
 tqdm
 ogb
 ray
-gpustat
-wandb
 torch
 torch-geometric>=2.3.0
