+++ pyproject.toml
@@ -18,7 +18,6 @@ requires-python = ">=3.8"
 
 dependencies = [
     "PyYAML>=5.4",
-    "requests>=2.22.0",
     "tensorboard>=2.11.0",
     "tqdm>=4.47.0",
     "ogb>=1.2.0",
