### 2023-09-09 #5713 @diningphil
Dear @idoby,

We are happy the changes are satisfactory. In order:

1. Thanks for spotting the (new) typo and the inconsistency! We fixed them.
2. We have removed wandb and gpustats from the dependencies as suggestes. If CUDA is used, an exception is now raised for gpustats as it already happened for wandb and the user is required to install the library.
3. **References:** most machine learning _conference_ papers do not have DOIs, whereas amongst journals the Journal of Machine Learning Research does not release a DOI. From the reference checker it seems the DOIs provided match the _journal_ papers that we referenced and no missing DOIs have been found among the others. This looks like correct behavior, but it is my first time with JOSS so I could well be wrong.

Thank you again for your careful review. We remain on stand-by for additional exchanges if required =).
