### 2023-09-09 #5713 @idoby
What? Every ML conference paper is supposed to publish a DOI.
For example, the DOI for Gilmer et el. Neural message passing for Quantum chemistry is:
https://dl.acm.org/doi/10.5555/3305381.3305512
