+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pydgn"
-version = "1.4.3"
+version = "1.4.4"
 description = "A Python Package for Deep Graph Networks"
 authors = [ { name="Federico Errica", email="f.errica@protonmail.com" } ]
 readme = "README.md"
@@ -18,6 +18,8 @@ requires-python = ">=3.8"
 
 dependencies = [
     "PyYAML>=5.4",
+    "Requests>=2.31.0",
+    "scikit_learn>=1.3.0",
     "tensorboard>=2.11.0",
     "tqdm>=4.47.0",
     "ogb>=1.2.0",
