### 2023-09-28 #5713 @sepandhaghighi
@arfon @diningphil 👋

I apologize for my delayed reply
Excellent work 🥇

Only two comments:

1. I advise incorporating a **PR template** into your repository
2. Since it appears that this library is only compatible with **Linux**, I believe it would be best to discuss it under the installation section (In my opinion, that Ubuntu badge is not clear enough)

SH
