### 2023-09-29 #5713 @diningphil
Dear @sepandhaghighi,

Thank you so much for your review! We have implemented the following changes:

1. Specify in Installation section (Readme and docs) that PyDGN works with linux only
2. Introduced a basic PR template as you recommended following [this link](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository).

If you are happy with our changes, I think we might ask @arfon to take the final decision.

Best regards,
the authors
