### 2023-09-29 #5713 @diningphil
> @diningphil What makes it Linux only?

Hi @idoby, PyDGN has never been tested on Linux, and the commands to prepare datasets and launch experiments assume a Linux distro.
