### 2023-09-29 #5713 @diningphil
@idoby wow, we really thank you for testing this yourself. We will try to test on Windows and see how it goes. If that is the case, we will remove the Ubuntu badge and then claim that it should work on most systems. We will keep you posted.

Thank you again for the help!!
