### 2023-09-29 #5713 @idoby
It's a best practice to set up a CI workflow to test every commit automatically on multiple operating systems and versions of Python. You should be able to find plenty of examples online, for example using GitHub Actions.

It's probably a good idea if you intend to continue developing this package.
Anyway, good luck! This paper can be published as far as I'm concerned. 🥳
