### 2023-09-29 #5713 @diningphil
I'll try to set it up directly with GitHub Actions as I did for Linux :)
