### 2023-09-30 #5713 @sepandhaghighi
>  @sepandhaghighi – could I ask that you check off any remaining review items in your checklist

@arfon Everything looks good to me 💯

> So @sepandhaghighi I will also remove ubuntu badge because it runs on any system apparently =)

@diningphil Good news 🔥

Please consider the following:

1. You should update `pyproject.toml`, `classifiers` section and add `Operating System :: OS Independent` instead of `Operating System :: POSIX :: Linux`

2. Including the minimum required version of Python in your document is recommended.

SH
