### 2023-09-30 #5713 @diningphil
> Please consider the following:
>
>     1. You should update `pyproject.toml`, `classifiers` section and add `Operating System :: OS Independent` instead of `Operating System :: POSIX :: Linux`
>
>     2. Including the minimum required version of Python in your document is recommended.

Both are done! Thank you for spotting them!
