### 2023-10-01 #5713 @diningphil
Dear @arfon,

- We applied small changes to the article (minor modification in abstract and acknowledgments to both reviewers. The rest looks ok.
- PyDGN release 1.5.0 now includes the changes suggested by the reviewers.
- This is the DOI generated with Zenodo: [https://doi.org/10.5281/zenodo.8396373](https://doi.org/10.5281/zenodo.8396373)

Please let me know if anything else is required.

Many thanks!
